# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HelixWeave API",
    "description": "Reproducible omics workflows, from inputs to evidence.",
    "version": "0.4.1"
  },
  "paths": {
...
    "/api/v1/runs/{run_id}/preflight": {
      "post": {
        "tags": [
          "preflight"
        ],
        "summary": "Trigger Preflight",
        "description": "Trigger local preflight for an existing run.\n\nThe actual planning/materialization/dry-run work runs in a background task\nso the HTTP response returns immediately with the run in ``VALIDATING``.",
        "operationId": "triggerPreflight",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InputBindingResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "compatibility_unresolved_v1",
              "declared_input_uses_v1"
            ],
            "title": "Mode"
          },
          "adapter_contract_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Adapter Contract Version"
          },
          "digest": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Digest"
          },
          "fully_managed": {
            "type": "boolean",
            "title": "Fully Managed"
          },
          "input_uses": {
            "items": {
              "$ref": "#/components/schemas/InputUseBindingResponse"
            },
            "type": "array",
            "maxItems": 256,
            "title": "Input Uses"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "digest",
          "fully_managed",
          "input_uses"
        ],
        "title": "InputBindingResponse",
        "description": "Allowlisted public projection of immutable input-use provenance."
      },
...
      "InputUseBindingResponse": {
        "properties": {
          "input_use_key": {
            "type": "string",
            "maxLength": 255,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,254}$",
            "title": "Input Use Key"
          },
          "occurrence": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Occurrence"
          },
          "capability_version": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Capability Version"
          },
          "closure_contract_version": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Closure Contract Version"
          },
          "provenance_mode": {
            "type": "string",
            "enum": [
              "managed_revision_v1",
              "transitional_unmanaged_v1"
            ],
            "title": "Provenance Mode"
          },
          "input_file_revision_ids": {
            "items": {
              "type": "string",
              "maxLength": 38,
              "pattern": "^inpfr_[0-9a-f]{32}$"
            },
            "type": "array",
            "maxItems": 256,
            "title": "Input File Revision Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "input_use_key",
          "occurrence",
          "capability_version",
          "closure_contract_version",
          "provenance_mode",
          "input_file_revision_ids"
        ],
        "title": "InputUseBindingResponse",
        "description": "Path-free public evidence for one adapter-owned input use."
      },
      "IssueResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "error"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "technical_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technical Message"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "IssueResponse",
        "description": "JSON-ready Issue shape from the PR84 contract."
      },
...
      "ReferenceProfileRevisionResponse": {
        "properties": {
          "profile_id": {
            "type": "string",
            "maxLength": 37,
            "pattern": "^refp_[0-9a-f]{32}$",
            "title": "Profile Id"
          },
          "revision_id": {
            "type": "string",
            "maxLength": 38,
            "pattern": "^refpr_[0-9a-f]{32}$",
            "title": "Revision Id"
          },
          "revision_number": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision Number"
          },
          "display_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Display Name"
          },
          "organism": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Organism"
          },
          "assembly": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Assembly"
          },
          "identity_sha256": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Identity Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "profile_id",
          "revision_id",
          "revision_number",
          "display_name",
          "organism",
          "assembly",
          "identity_sha256"
        ],
        "title": "ReferenceProfileRevisionResponse",
        "description": "Allowlisted public identity for one immutable reference revision."
      },
...
      "RunRecordResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "cancellation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Reason"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IssueResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tags"
          },
          "reference_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReferenceProfileRevisionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "workflow_id",
          "status",
          "created_at",
          "updated_at",
          "started_at",
          "ended_at",
          "current_stage",
          "cancellation_reason"
        ],
        "title": "RunRecordResponse",
        "description": "Allowlisted public projection of one internal RunRecord."
      },
      "RunResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunRecordResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "RunResponse",
        "description": "Envelope for single-run endpoints."
      },
...
      "ValidatedInputSnapshotResponse": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "maxLength": 38,
            "pattern": "^vsnap_[0-9a-f]{32}$",
            "title": "Snapshot Id"
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "adapter_version": {
            "type": "string",
            "title": "Adapter Version"
          },
          "payload_digest": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Payload Digest"
          },
          "validated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Validated At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "project_id": {
            "type": "string",
            "maxLength": 36,
            "pattern": "^prj_[0-9a-f]{32}$",
            "title": "Project Id"
          },
          "binding_mode": {
            "type": "string",
            "enum": [
              "legacy_v1",
              "bound_v1"
            ],
            "title": "Binding Mode"
          },
          "provenance": {
            "type": "string",
            "enum": [
              "resolved",
              "unresolved"
            ],
            "title": "Provenance"
          },
          "sample_revision_ids": {
            "items": {
              "type": "string",
              "maxLength": 37,
              "pattern": "^smpr_[0-9a-f]{32}$"
            },
            "type": "array",
            "title": "Sample Revision Ids"
          },
          "binding_digest": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Binding Digest"
          },
          "input_binding": {
            "$ref": "#/components/schemas/InputBindingResponse"
          },
          "reference_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReferenceProfileRevisionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "snapshot_id",
          "workflow_id",
          "schema_version",
          "adapter_version",
          "payload_digest",
          "validated_at",
          "expires_at",
          "project_id",
          "binding_mode",
          "provenance",
          "sample_revision_ids",
          "binding_digest",
          "input_binding"
        ],
        "title": "ValidatedInputSnapshotResponse",
        "description": "Safe public projection of one server-owned validation snapshot."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ValidationResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          },
          "value": {
            "type": "null",
            "title": "Value"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidatedInputSnapshotResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "workflow_id",
          "snapshot"
        ],
        "title": "ValidationResponse",
        "description": "Envelope for POST /api/v1/workflows/{workflow_id}/validate."
      },
...
    }
  }
}